add update method to SearchBar component
