+++ client/public/assets/css/search.css
@@ -9,6 +9,7 @@
   /* border: 1px solid red; */
   width: 25%;
   height: 5vh;
+  max-height: 60px;
   display: flex;
   align-items: center;
   justify-content: space-between;
