+++ client/src/components/App.js
@@ -5,12 +5,12 @@ import { Route, Switch, Link } from "react-router-dom";
 import Login from "./auth/Login";
 import GenreIndex from "./genres/GenreIndex";
 import AuthRoute from "../util/route_util";
-import Nav from "./Nav";
+import Nav from "./navigation/Nav";
 import Register from "./auth/Register";
-import SideBar from './SideBar';
+import SideBar from './navigation/SideBar';
 import MusicPlayer from "./player/MusicPlayer";
-import Search from './Search';
-import Account from './Account';
+import Search from './navigation/Search';
+import Account from './navigation/Account';
 import SplashPage from "./Splash";
