+++ client/src/components/navigation/Nav.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Link, withRouter } from "react-router-dom";
 import { Query, ApolloConsumer } from "react-apollo";
-import Queries from "../graphql/queries";
+import Queries from "../../graphql/queries";
 const { IS_LOGGED_IN } = Queries;
