+++ client/src/graphql/mutations.js
@@ -24,5 +24,6 @@ export default {
         loggedIn
       }
     }
-  `
+  `,
+  
 }
